% Auto-generated: do not edit by hand
\name{''DashBootstrapDaterangepicker}

\alias{''DashBootstrapDaterangepicker}

\title{DashBootstrapDaterangepicker component}

\description{
DashBootstrapDaterangepicker component based on react-bootstrap-daterangepicker
}

\usage{
''DashBootstrapDaterangepicker(id=NULL, className=NULL, end_date=NULL,
initialSettings=NULL, innerClassName=NULL, start_date=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Class of the input component.}

\item{end_date}{Character. end_date}

\item{initialSettings}{Named list. You can pass all the settings from the original plugin (http://www.daterangepicker.com/)
to the initialSettings prop:

<input>, alwaysShowCalendars, applyButtonClasses, applyClass, autoApply, autoUpdateInput, buttonClasses,
cancelButtonClasses, cancelClass, dateLimit, drops, endDate, isCustomDate, isInvalidDate, linkedCalendars,
locale, maxDate, maxSpan, maxYear, minDate, minYear, moment, opens, parentEl, ranges, showCustomRangeLabel,
showDropdowns, showISOWeekNumbers, showWeekNumbers, singleDatePicker, startDate, template, timePicker,
timePicker24Hour, timePickerIncrement, timePickerSeconds}

\item{innerClassName}{Character. Class of the input.}

\item{start_date}{Character. start_date}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

